pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ui-cli"
version = "0.1.0"
description = "CLI utility for UniFi Site Manager API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Vedanta", email = "vedanta@example.com" }
]
keywords = ["unifi", "ubiquiti", "cli", "network", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
]
dependencies = [
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
ui = "ui_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ui_cli"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require real API)",
    "slow: marks tests as slow",
]
addopts = "-v --tb=short"
